[Feature] Refactor, Matched Jon's Syntax
