+++ server/controller/search.getRestaurantList.js
@@ -1,0 +1,16 @@
+var searchGooglePlacesModel = require('../model/searchGooglePlaces.model.js');
+
+exports.searchGooglePlaces = {
+	getRestaurantList: getRestaurantList
+}
+
+function getRestaurantList(req, res) {
+	searchGooglePlacesModel.getRestaurantList(req.body)
+	.then(function(result) {
+		res.status(200).send(result)
+	})
+	.catch(function(err) {
+		res.status(500).end('Error in getRestaurantList', err)
+	})
+}
+
